requirement.txt
streamlit
pandas
numpy
plotly
requests
fastapi
uvicorn
pdfplumber
pydantic
google-generativeai
sentence-transformers
chromadb
matplotlib
